requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
aiohttp==3.9.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
bcrypt==4.0.1

# AI and ML dependencies
numpy==1.24.3
scikit-learn==1.3.2

# Configuration and utilities
PyYAML==6.0.1
tenacity==8.2.3
structlog==23.2.0
prometheus-client==0.19.0
redis==5.0.1

# Monitoring and health checks
httpx==0.25.2
prometheus-fastapi-instrumentator==6.1.0
psutil==5.9.7
py-healthcheck==1.10.1

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Documentation
fastapi-versioning==0.10.0
fastapi-utils==0.2.1
openapi-spec-validator==0.7.1